% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidish.R
\name{get_epidish_signature_matrix}
\alias{get_epidish_signature_matrix}
\title{Get EpiDISH Signature Matrix}
\usage{
get_epidish_signature_matrix(reference = c("blood", "breast", "epithelial"))
}
\arguments{
\item{reference}{One of 'blood', 'breast', 'epithelial'.}
}
\value{
Signature matrix as tibble with CpGs in rows (column 'CpGs') and cell types in columns
}
\description{
Returns the reference matrix used by EpiDISH for a given type.
}
