% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/houseman.R
\name{get_houseman_signature_matrix}
\alias{get_houseman_signature_matrix}
\title{Get Houseman Signature Matrix}
\usage{
get_houseman_signature_matrix()
}
\value{
Signature matrix as tibble with CpGs in rows (column 'CpGs') and cell types in columns
}
\description{
Returns the IDOLOptimizedCpGs.compTable (signature matrix) used by Houseman for a given array type..
}
