% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methatlas.R
\name{get_methatlas_signature_matrix}
\alias{get_methatlas_signature_matrix}
\title{Get MethAtlas Signature Matrix}
\usage{
get_methatlas_signature_matrix(
  reference_atlas = system.file("reference_atlas.csv", package = "methyldeconv")
)
}
\arguments{
\item{reference_atlas}{Path to a csv file (default: package reference_atlas.csv)}
}
\value{
Signature matrix as tibble with CpGs in rows (column 'CpGs') and cell types in columns
}
\description{
Returns the reference matrix used by MethAtlas from a CSV file.
}
