% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylcc.R
\name{get_methylcc_signature_matrix}
\alias{get_methylcc_signature_matrix}
\title{Get MethylCC Signature Matrix (DMRs)}
\usage{
get_methylcc_signature_matrix()
}
\value{
DMRs as dataframe (as used by methylCC)
}
\description{
Runs find_dmrs on the default reference and returns the DMRs.
}
