% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylResolver.R
\name{get_methylresolver_signature_matrix}
\alias{get_methylresolver_signature_matrix}
\title{Get MethylResolver Signature Matrix}
\usage{
get_methylresolver_signature_matrix()
}
\value{
Signature matrix as tibble with CpGs in rows (column 'CpGs') and cell types in columns
}
\description{
Returns the signature matrix used by MethylResolver.
}
